requirements.txt
fastapi # para crear la API web
uvicorn # para el servidor ASGI
sqlalchemy # para manejar la base de datos
pydantic  # para validación de datos
bcchapi # para interactuar con la API de la Bolsa de Comercio de Chile
numpy # para manejar datos numéricos
transbank # para integrar con Transbank
python-dotenv # para cargar variables de entorno desde un archivo .env